requirements.txt
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
et-xmlfile==1.0.1
idna==2.10
iniconfig==1.0.1
jdcal==1.4.1
lxml==4.5.2
MarkupPy==1.14
numpy==1.19.2
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.4
pandas==1.1.3
pandas-datareader==0.9.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
tablib==2.0.0
tabulate==0.8.7
toml==0.10.1
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
